Move controller JS and template files
Step 1 of the instructions for converting to the new component style
says to move these two files to the /component/modal directory.
This makes them a colocated component. The /component/modal directory
didn't currently exist, so I created it and moved the files
there.

I referenced the following example to know where to move these files:
https://github.com/discourse/discourse/pull/22154/files#diff-f7172861de6b7fc4cf2365b1296f1e90d828c91a4393ae9714e9ea1ebf7b06d0