Change to check for valid class attributes in DOM
This acceptance test was opening up the options when creating a
topic, and trying to check that there was an attribute named
data-value that has the value showQuizBuilder. However, in DOM,
there's no attribute for the quiz builder named data-value. I've
changed it to check for the attribute data-name, and that there's
one with the value 'Build Quiz'.